refactor pm quality summary invocation preview router
